requirements.txt
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
falcon==1.4.1
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
pbr==4.0.2
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.1
pytest==3.5.1
pytest-cov==2.5.1
python-dateutil==2.7.2
python-mimeparse==1.6.0
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
